WIP: Respect OUTPUT_DIRECTORY properties (CMake >= 3.19)
Respect the following `OUTPUT_DIRECTORY` target properties if set:
-  ARCHIVE_OUTPUT_DIRECTORY
-  LIBRARY_OUTPUT_DIRECTORY
-  RUNTIME_OUTPUT_DIRECTORY
-  PDB_OUTPUT_DIRECTORY

For Multi-Config generators the respective `_CONFIG` variants are
also respected.

The build artifacts are copied into the expected location.
If the target properties are not set or the feature toggle is set
to OFF, fall back to ${CMAKE_CURRENT_BINARY_DIR}.